+++ WebApplication/src/orthancApi.js
@@ -220,27 +220,6 @@ export default {
         return (await axios.get(orthancApiUrl + "tools/log-level")).data;
     },
 
-    async shareStudy(orthancId, dicomId, anonymized, expirationInDays) {
-        let expirationDate = null;
-        
-        if (expirationInDays > 0) {
-            let today = new Date();
-            expirationDate = new Date();
-            expirationDate.setDate(today.getDate() + expirationInDays);
-        }
-                
-        const response = (await axios.put(oe2ApiUrl + "shares", {
-            "studies" : [{
-                "dicom-uid": dicomId,
-                "orthanc-id" : orthancId
-            }],
-            "anonymized": anonymized,
-            "expiration-date": expirationDate
-        }));
-        
-        return response.data['url'];
-    },
-
     async getMedDreamInstantLink(dicomId) {
         const response = (await axios.put(oe2ApiUrl + "shares", {
             "studies" : [{